Check for empty translation
GP doesn't let a user to submit an empty translation. However, it doesn't reject empty plurals when importing.

So this adds a check to `wpgpt_check_all_translations()` as `wpgpt_check_this_translation()` will never have on submission an empty translation.

Adds warning icon, a warning in the editor list and a warning label that doesn't respect user settings on purpose!