Highlights some hard to spot warnings
The following warnings/notices are highlighted
- double spaces
- warning and notice user defined words

It doesn't highlight placeholders as it doesn't know how to pick from a string with multiple placeholders in a random order.

It doesn't highlight additional start/end space/period symbol/column/etc. as the label is pretty explanatory itself, no need to over colorize the string.

When a highlight is present, the corresponding label is hidden.

The highlight has the same colour regardless of the warning or notice state of the check - known behaviour.

The highlight happens only if
- setting "Warning labels and highlights" is enabled
- the check in case is set as Notice or Warning