Count and match strings
Adds an option to set a list of strings that will be counted in original and translation and then matched. A warning will be generated if counts don't match.

Replaces regex in checks with optimized function, where possible. Up to 10 times faster, according to author.

Removes the break in finding the warning words after the first word is found.

Removes the "Notify about these words" option.

Fixes:
https://github.com/vlad-timotei/wpgp-tools/issues/14,
https://github.com/vlad-timotei/wpgp-tools/issues/19,
https://github.com/vlad-timotei/wpgp-tools/issues/8