Feature | Consistency: Anonymous user
After https://meta.trac.wordpress.org/changeset/11154 we are able to query for anonymous but this lacks an UI.

This implements an idea of @Presskopp to show a checkbox that autocompletes the above field with 'anonymous'.

Note: This commit is here to give proper props and keep track of things, but the actual code for this feature is in Enhancement | Consistency: refactoring commit. :)